+++ src/main/java/com/lambdaschool/countries/NamesController.java
@@ -21,14 +21,14 @@ public class NamesController {
     @GetMapping(value = "/start/{letter}", produces = {"application/json"})
     public ResponseEntity<?> getCountriesAlphabeticallyBeginningWith(@PathVariable char letter) {
         ArrayList<Country> rtnCountries = CountriesApplication.ourCountryList.findCountries(c -> c.getName().toUpperCase().charAt(0) == Character.toUpperCase(letter));
-        rtnCountries.sort((o1, o2) -> o1.getName().compareToIgnoreCase(o2.getName()));
+        rtnCountries.sort((o1, o2) -> (int)(o1.getName().compareToIgnoreCase(o2.getName())));
         return new ResponseEntity<>(rtnCountries, HttpStatus.OK);
     }
 
     @GetMapping(value = "/size/{number}", produces = {"application/json"})
     public ResponseEntity<?> getCountriesWithNamesGreaterThan(@PathVariable int number) {
         ArrayList<Country> rtnCountries = CountriesApplication.ourCountryList.findCountries(c -> c.getName().length() > number);
-        rtnCountries.sort((o1, o2) -> o1.getName().compareToIgnoreCase(o2.getName()));
+        rtnCountries.sort((o1, o2) -> (int)(o1.getName().compareToIgnoreCase(o2.getName())));
         return new ResponseEntity<>(rtnCountries, HttpStatus.OK);
     }
 }
